+++ OniAccess/OniAccessStrings.cs
@@ -85,6 +85,7 @@ namespace STRINGS {
 			public static LocString FLOW_SONIFICATION = "Flow sonification";
 			public static LocString TEMPERATURE_BAND_EARCONS = "Temperature band earcons";
 			public static LocString FOLLOW_MOVEMENT_EARCONS = "Follow movement earcons";
+			public static LocString FOOTSTEP_EARCONS = "Footstep earcons";
 		}
 
 		// Supply closet (Klei rewards) screen messages